Project edit
First edit to code, got rid of extra material I was intending to study so It looks more professional and neat. Moved some things around like the tap, made the background image cover more than half of the screen and also increased size of bottle and stand to allow user to potentially be able to click on tap. Also made it bigger for devices like Ipad air and pro.